Add variable highlight feature
This feature continuously highlights all occurences of the variable at
the point in the same statement. This is useful for identifying the
way the variable is defined in the statement and spotting typos.

The code now immediatly highlights the variable at the point, but it
might be better to introduce a wait period as used in
highlight-symbol.el to reduce the visual clutter and restlessness
caused by flickering colors when fastly moving through code.